Improvements to target detection for the gear placement.
This detects vertically oriented rectangles of a minimum size
and screens out other clutter in the image.  It highlights
the rectangle in an imshow window in blue...the rectangles might
be rotated, so it shows just vertical orientations in red (i.e.,
trying to make a rectangle from just two of the four points we're
given, which, after experimentation, turns out to not be a very
good idea).

Note that this code was written to work on my Mac, where I can only
get imshow windows working against OpenCV 2.4.  There are a few
compatibility issues with OpenCV 3, so this probably won't work right
out of the box with 3 (for example, the number of arguments returned
by findContours is different).