Changes from Mr. Fultz re image processing thresholds.
With the new red-mask processing, we can now make the kernel
for the open/close operations smaller because there's not as much
worry about clutter in real-world images.  And tweaked the
luminescence threshold a bit more.

Also, flipped the order of the opening and closing operation because
that seems to work to prevent the rectangles from disappearing
inappropriately.