Reenable imshows for debugging... -- Mr. Fultz
Also, it turns out that cv2.findContours() changes the image you pass into it.  Which
confused me for a while when I was trying to figure out why the image processing
behaved as it did.  In order to compensate, I just moved the imshow for that
data to above the findContours call.